fix: profile/posts 페이지 Suspense 적용
